[ALPS07688835] wlan: Fix code defect
Fix code defect

MTK-Commit-Id: dc1bdd6b8b3ce33a20efc0e17de7d82b1ce94bed

CR-Id: ALPS07688835
Feature: Wi-Fi Chipset Capability
(cherry picked from commit e4d9dff9f251afca322ce718f6760e46573d454b)
(cherry picked from commit 889818cbff76bd5d07c01502dacb40b6c4a59a31)
(cherry picked from commit 814362a4145cde74c7d55ac14e87f55b91140e8d)
(cherry picked from commit 0949e741b9ca58d1f94439e03388068ba3b96bb2)
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/7553702
Coverity-Review-Label: srv_pf_nep_win <srv_pf_nep_win@mediatek.com>
Build: srv_preflight_b001 <srv_preflight_b001@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_pf_nep_win <srv_pf_nep_win@mediatek.com>
Commit-Check: srv_pis <srv_pis@mediatek.com>
Test: srv_pf_nep_sanity
Test: srv_mspautosanity
Coverity-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Tested-by: chris.hong <chris.hong@mediatek.com>

+++ os/linux/gl_kal.c
@@ -7325,9 +7325,9 @@ int32_t kalRequestFirmware(const uint8_t *pucPath,
 	       pucPath);
 
 	if (ucIsZeroPadding)
-		u4Size = fw->size + 1;
+		u4Size = (uint32_t) fw->size + 1;
 	else
-		u4Size = fw->size;
+		u4Size = (uint32_t) fw->size;
 
 	pucData = kalMemAlloc(u4Size, VIR_MEM_TYPE);
 	if (pucData == NULL) {
